✨ (feat): Re-add partial upload
This commit readds partial upload.
It uses a cloudflare endpoint instead of a supabase one
It sends the pre-signed URL request in batches of 10 making it less efficient than before but more stable
A new 'cloudflareBaseUrl' option has been added to the capacitor config. This option allows to control the baseUrl for partial upload
ZOD validation has beed added compared to old partial upload. This decreases the chance of bugs occurring